osg-post-install: Split out staging dir from final OSG_LOCATION (SOFTWARE-1126); cleanup some code
In order to support installation into OASIS, the osg-post-install script needs
to distinguish between the staging directory (the dir that the tarball was
extracted into) and the final OSG_LOCATION (the dir that should be embedded
into the config and environment files).

Added optparse to accomplish this; there's now a --final-osg-location (-f)
argument that can be passed to use a different dir for the final OSG_LOCATION.
The first argument has been renamed to STAGING_DIR, but will be reused for
OSG_LOCATION if the latter is not specified. Updated the usage to mention that.

git-svn-id: https://vdt.cs.wisc.edu/svn/software/tarball-client/trunk@17305 4e558342-562e-0410-864c-e07659590f8c